requirements.txt
selenium
requests
python-dotenv
beautifulsoup4
